Add DynamoDB payment persistance

+++ lambda/processors/payments.py
@@ -1,4 +1,5 @@
 from models.payment import Payment
+from decimal import Decimal
 
 
 def parse_payment_row(row: str) -> Payment:
@@ -6,7 +7,7 @@ def parse_payment_row(row: str) -> Payment:
     return Payment(
         payment_id=int(values[0]),
         client=values[1].strip(),
-        amount=float(values[2]),
+        amount=Decimal(values[2]),
         date=values[3],
         status=values[4].lower(),
     )
